requirements.txt
streamlit>=1.20
Pillow>=9.0
numpy
pandas
plotly
seaborn
torch; extra == 'torch'
torchvision; extra == 'torchvision'
